move update update user hooks to shared
